requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.25.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg>=3.1.12
psycopg-binary>=3.1.12
SQLAlchemy[psycopg]>=2.0.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication & Security
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Caching
redis>=5.0.0
aioredis>=2.0.0

# Testing
pytest>=7.0.0
pytest-env>=1.1.0
pytest-dotenv>=0.5.2
pytest-asyncio>=0.23.0
fakeredis>=2.0.0
httpx>=0.26.0

# Email Validation
email-validator>=2.0.0 